style: 패키지 네이밍 컨벤션 수정
기존 카멜 케이스에서 모두 소문자를 사용하는 것으로 수정
