--- src/main/java/com/app/domain/memberSavedProblem/repository/MemberSavedProblemRepository.java
@@ -1,7 +1,0 @@
-package com.app.domain.memberSavedProblem.repository;
-
-import com.app.domain.memberSavedProblem.entity.MemberSavedProblem;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberSavedProblemRepository extends JpaRepository<MemberSavedProblem, Long> {
-}
