--- src/main/java/com/app/domain/memberSavedSummary/repository/MemberSavedSummaryRepository.java
@@ -1,7 +1,0 @@
-package com.app.domain.memberSavedSummary.repository;
-
-import com.app.domain.memberSavedSummary.entity.MemberSavedSummary;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberSavedSummaryRepository extends JpaRepository<MemberSavedSummary, Long> {
-}
