+++ src/main/java/com/app/domain/membersavedproblem/repository/MemberSavedProblemRepository.java
@@ -1,0 +1,7 @@
+package com.app.domain.membersavedproblem.repository;
+
+import com.app.domain.membersavedproblem.entity.MemberSavedProblem;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberSavedProblemRepository extends JpaRepository<MemberSavedProblem, Long> {
+}
